refactor: move transport functions into utils package
