+++ pkg/utils/transport.go
@@ -1,4 +1,4 @@
-package operator2
+package utils
 
 import (
 	"crypto/tls"
@@ -12,8 +12,8 @@ import (
 
 // TODO move all this to library-go
 
-// transportFor returns an http.Transport for the given ca and client cert data (which may be empty)
-func transportFor(serverName string, caData, certData, keyData []byte) (http.RoundTripper, error) {
+// TransportFor returns an http.Transport for the given ca and client cert data (which may be empty)
+func TransportFor(serverName string, caData, certData, keyData []byte) (http.RoundTripper, error) {
 	transport, err := transportForInner(serverName, caData, certData, keyData)
 	if err != nil {
 		return nil, err
